+++ galacteek/ld/rdf/sync/cfg.py
@@ -2,6 +2,16 @@ import attr
 
 
 @attr.s(auto_attribs=True)
+class GraphSyncConfig:
+    # URI of graph synchronizer
+    use: str
+
+    hbPeriodicSend: bool = False
+    hbIntervalMin: int = 60
+    hbIntervalMax: int = 120
+
+
+@attr.s(auto_attribs=True)
 class GraphExportSyncConfig:
     type: str = 'rdfexport'
     format: str = 'ttl'
@@ -21,4 +31,4 @@ class GraphSemChainSyncConfig:
     recordsPerSync: int = 256
     recordFetchTimeout: int = 30
     syncIntervalMin: int = 60
-    chainSyncIntervalMin: int = 120
+    chainSyncIntervalMin: int = 90